### 2021-05-05 #114 @whedon
[PRE REVIEW]: ApplNumComp: An Open Access Introductory Course for Applied Numerical Computing
**Submitting author:** @ashleefv (<a href="http://orcid.org/0000-0001-9059-5703">Ashlee N. Ford Versypt</a>)
**Repository:** <a href="https://github.com/ashleefv/ApplNumComp" target ="_blank">https://github.com/ashleefv/ApplNumComp</a>
**Version:** v1.1

@ashleefv if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
