+++ paper.bib
@@ -3,7 +3,8 @@
   TITLE = {An interdisciplinary elective course to build computational skills for mathematical modeling in science and engineering},
   YEAR = {2019},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
-  url = {https://doi.org/10.18260/1-2--32072},
+  doi = {10.18260/1-2-{}-32072},
+  url = {https://doi.org/10.18260/1-2-{}-32072},
   City= {Tampa, FL}
 }
 
