### 2021-05-05 #114 @ashleefv
> 👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/jose-papers/jose.00114/jose.00114/10.21105.jose.00114.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/jose-papers/blob/jose.00114/jose.00114/10.21105.jose.00114.pdf) 📄 👈

proof looks good, except that my last name is Ford Versypt and so the citation at the bottom of the pages should be Ford Versypt et al.
