### 2021-05-06 #114 @ThomasA
I can contribute if needed. I have many years' experience with LaTeX, Git, Python, and Matlab; both using and teaching.
