### 2021-05-23 #114 @labarba
@ashleefv — I am not quite satisfied with the _paper_, as it relates to the Statement of Need, and to how the materials might be used by others. Note this from the [submission requirements](https://openjournals.readthedocs.io/en/jose/submitting.html#submission-requirements):

> Computational learning modules should be complete and immediately usable for self-learning or adoption by other instructors.

And also (further below):

> The goal is that someone reading the JOSE paper has enough information to decide if they’d be interested in adoping the learning module or software.

From our [scope](https://jose.theoj.org/about) on the website, note that the JOSE review is centered on the text-based open-source materials, and we exclude slides, lecture notes, or YouTube videos—these can be considered supplementary materials, if present, but are not the focus of the review.
