+++ paper.md
@@ -33,7 +33,7 @@ ApplNumComp is a repository of open educational resources supporting an introduc
 # Statement of Need
 An introductory course on Applied Numerical Computing was developed for senior undergraduate and graduate students interested in research and industrial projects involving mathematical modeling using computational tools. The class was designed to fill a gap distinct from introductory programming and numerical methods courses and was intended to be very practical. Graduate students new to computational research were the primary student group targeted for the course topics. We have previously detailed the course [@FordVersypt2019] and the final course project [@Ruggiero2018].
 
-No single textbook covers all of the topics in the Applied Numerical Computing course. As the course emphasized best practices for reproducible research computing, we sought to apply the concept to the course materials for a reproducible computing educational experience. Here, we have curated a set of open educational resources for others to engage with the course lessons outside of the institution's learning management system and outside of the formal course. The main repository is organized into a set of lessons that include videos, activities, related references, and other instructional materials.  
+No single textbook covers all of the topics in the Applied Numerical Computing course. As the course emphasized best practices for reproducible research computing, we sought to apply the concept to the course materials for a reproducible computing educational experience. Here, we have curated a set of open educational resources for others to engage with the course lessons outside of the institution's learning management system and outside of the formal course. The main repository is organized into a set of lessons that include videos, activities, related references, and other instructional materials. Each lesson can be considered as a stand alone module on a particular computaitonal topic, but there are conceptual links between lessons. Many of the lessons include detailed MATLAB or Python coding examples. For instance, Lesson 13: Parameter Estimation in Python provides a worked example for fitting parameters to one differential equation and the solution .py file followed by a second example for fitting parameters to multiple differential equations. Documentation associated with MATLAB or Python often only shows much simpler examples. Related Lessons 11 and 12 provide supportive and complimentary content for general information and for examples in MATLAB. The full list of lessons is available [here](https://github.com/ashleefv/ApplNumComp#lessons). Other supporting materials include a list of suggested readings, the full set of computational assignments, and the course learning objectives.
  
 # Author Contributions
 Ashlee N. Ford Versypt developed the Applied Numerical Computing course at Oklahoma State University and taught it annually 2016-2020. The materials presented here were developed and delivered in a hybrid of in-person, online synchronous, and online asynchronous formats in Fall 2020. Duncan H. Mullins, a PhD student from Dr. Ford Versypt’s lab who took the course in Fall 2019, served as a teaching assistant for the course and aided in conversion of the course to this online open access version.
