### 2021-06-23 28ade4a
Update paper.md
expanded statement of need
