+++ paper.md
@@ -28,15 +28,18 @@ bibliography: paper.bib
 ---
 
 # Summary
-ApplNumComp is a repository of open educational resources supporting an introductory course on Applied Numerical Computing. Topics include Git for version control, LaTeX for typesetting, reproducible research computing, and MATLAB and Python for high-level programming and scientific computing applications of solving systems of differential equations, estimating parameters for models using regression, creating publication quality figures, and developing graphical user interfaces. The lessons can be used collectively for a similar course at another institution, for independent study by an online learner, or for incorporation of individual lessons into other courses that involve related topics. 
+ApplNumComp is a repository of open educational resources supporting an introductory course on Applied Numerical Computing. Topics include Git for version control, LaTeX for typesetting, reproducible research computing, and MATLAB and Python for high-level programming and scientific computing applications of solving systems of differential equations, estimating parameters for models using regression, creating publication quality figures, and developing graphical user interfaces. The lessons can be used collectively for a similar course at another institution, for independent study by an online learner, or for incorporation of individual lessons (modules or examples) into other courses that involve related topics. 
  
 # Statement of Need
-An introductory course on Applied Numerical Computing was developed for senior undergraduate and graduate students interested in research and industrial projects involving mathematical modeling using computational tools. The class was designed to fill a gap distinct from introductory programming and numerical methods courses and was intended to be very practical. Graduate students new to computational research were the primary student group targeted for the course topics. We have previously detailed the course [@FordVersypt2019] and the final course project [@Ruggiero2018].
+An introductory course on Applied Numerical Computing was developed for upper division undergraduate and graduate students interested in research and industrial projects involving mathematical modeling using computational tools. The class was designed to fill a gap distinct from introductory programming and numerical methods courses and was intended to be very practical. Graduate students new to computational research were the primary student group targeted for the course topics. We have previously detailed the course [@FordVersypt2019] and the final course project [@Ruggiero2018].
 
-No single textbook covers all of the topics in the Applied Numerical Computing course. As the course emphasized best practices for reproducible research computing, we sought to apply the concept to the course materials for a reproducible computing educational experience. Here, we have curated a set of open educational resources for others to engage with the course lessons outside of the institution's learning management system and outside of the formal course. The main repository is organized into a set of lessons that include videos, activities, related references, and other instructional materials. Each lesson can be considered as a stand alone module on a particular computaitonal topic, but there are conceptual links between lessons. Many of the lessons include detailed MATLAB or Python coding examples. For instance, Lesson 13: Parameter Estimation in Python provides a worked example for fitting parameters to one differential equation and the solution .py file followed by a second example for fitting parameters to multiple differential equations. Documentation associated with MATLAB or Python often only shows much simpler examples. Related Lessons 11 and 12 provide supportive and complimentary content for general information and for examples in MATLAB. The full list of lessons is available [here](https://github.com/ashleefv/ApplNumComp#lessons). Other supporting materials include a list of suggested readings, the full set of computational assignments, and the course learning objectives.
[…]
