### 2021-06-23 5452308
Update paper.md
Further expansion of statement of need section
