### 2021-06-23 #114 @ashleefv
Sorry for the delay in following up. It's been a busy summer already. @labarba Based on your feedback, I have made a substantial extension of paper.md by expanding the 2nd paragraph of the statement of need and adding the 3rd paragraph on our experiences with the course and the materials.
