### 2021-07-28 #114 @labarba
@moorepants – you can mark yourself as Out of Office using the `/ooo` bot.
See https://openjournals.readthedocs.io/en/jose/editing.html#out-of-office
