### 2021-08-02 #114 @sconde
> Hello @sconde, are you interested in (and available to) review this submission for JOSE?

yes.
