### 2021-10-11 #143 @whedon
[REVIEW]: ApplNumComp: An Open Access Introductory Course for Applied Numerical Computing
**Submitting author:** @ashleefv (<a href="http://orcid.org/0000-0001-9059-5703">Ashlee N. Ford Versypt</a>)
**Repository:** https://github.com/ashleefv/ApplNumComp
**Branch with paper.md** (empty if default branch):
**Version:** v2.0

- [ ] **Version:** Does the release version given match the repository release (v1.1)?

- [x] **Version:** Does the release version given match the repository release (v1.1)?
