### 2021-11-02 #143 @labarba
Hi everyone! Could we have a little update? When do you think you could tick off a few more items of your checklist, @sconde @ThomasA ?
