### 2021-11-28 ashleefv/ApplNumComp#2 @sconde
JOSE-Review-143: Statement of need
> [ ] **A statement of need:** Do the authors clearly state the need for this module and who the target audience is?

Although there is a `Statement of Need` section in the article, the section does not clarify the need for this material to the intended audiences.

It would help to give some examples to support the following statement:
> The class was designed to fill a gap distinct from introductory programming and numerical methods courses and was
intended to be very practical.
