### 2021-11-28 #143 @sconde
> [ ] **A statement of need:** Do the authors clearly state the need for this module and who the target audience is?

https://github.com/ashleefv/ApplNumComp/issues/2
