### 2021-11-28 ashleefv/ApplNumComp#3 @sconde
JOSE-Review-143: instalation instructions
> * [ ]  **Installation instructions:** Is there a clearly stated list of dependencies?

The course is using the overleaf online editor for the latex material and that is a terrific idea to introduce someone to latex.

The course is using MATLAB and Python. However, there are no instructions for installing MATLAB and Python.

It's great the course provides instruction videos that students and others can follow along, but it would be ideal to provide a Jupyter notebook for the python materials and a MATLAB live script.
