### 2021-11-28 #143 @sconde
> * [ ]  **Installation instructions:** Is there a clearly stated list of dependencies?
> *
https://github.com/ashleefv/ApplNumComp/issues/3
