### 2021-11-28 ashleefv/ApplNumComp#4 @sconde
JOSE-Review-143: Missing community guidelines
The repository is missing a `Contributing, Issue Reporting, and Feature Requests` section
