### 2021-11-28 #143 @sconde
> * **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

https://github.com/ashleefv/ApplNumComp/issues/4
