### 2022-06-08 7bac99d
Update L13 Parameter Estimation in Python.md
included a link to extended example jupyter notebook
