+++ L13 Parameter Estimation in Python.md
@@ -200,6 +200,9 @@ plt.show()
 ![Expected Graph 2](/Lesson_images/Figure2_L13.png)
 * Solution [.py file](/CHEclassFa20/In%20Class%20Problem%20Solutions/Python/ODEParamEstimExample2.py)
 
+## **Extended Example**
+* [Parameter estimation of ODE models for chemical kinetics](/CHEclassFa20/In%20Class%20Problem%20Solutions/Python/ParameterEstimationOfODEModels.ipynb)
+
 ## **Reference for Further Exploration**
 * [Python curve fitting tools](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.curve_fit.html)
 
