### 2022-06-21 a19d3bb
Update RecommendedReading.md
Added a new reference for reproducible research computing that was published after the original list was compiled
