+++ README.md
@@ -2,7 +2,7 @@
 
 This repository contains a set of lessons on Applied Numerical Computing covering Git for version control, LaTeX for typesetting, and MATLAB and Python for high-level programming and scientific computing. 
 
-Note: this site is based on the Fall 2020 course offering: CHE 4753/5753 Applied Numerical Computing for Scientists & Engineers at Oklahoma State University created and taught by Ashlee N. Ford Versypt, Ph.D. and assisted by Duncan H. Mullins. The translation of course materials to the online lessons here was supported by a mini-grant from the [Computer Aids in Chemical Engineering (CACHE) Corporation](https://cache.org/). This material is based upon work supported by the National Science Foundation under Grant No. 1845117 to Dr. Ford Versypt. 
+Note: this site is based on the Fall 2020 course offering: CHE 4753/5753 Applied Numerical Computing for Scientists & Engineers at Oklahoma State University created and taught by Ashlee N. Ford Versypt, Ph.D. and assisted by Duncan H. Mullins. The translation of course materials to the online lessons here was supported by a mini-grant from the [Computer Aids in Chemical Engineering (CACHE) Corporation](https://cache.org/). This material is based upon work supported by the National Science Foundation under Grant No. 2133411 (formerly 1845117) to Dr. Ford Versypt. 
 
 Course repository:
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4711792.svg)](https://doi.org/10.5281/zenodo.4711792)
