### 2022-12-30 #143 @sconde
> > * **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

@ashleefv This issue still needs to be resolved.
