### 2022-12-30 #143 @sconde
> * [x]  **Version:** Does the release version given match the repository release (v1.1)?
> *

@ashleefv, does the release version need to be updated to reflect the changes you've made?
