### 2022-12-30 #143 @kyleniemeyer
@ashleefv I realize that I haven't done a great job shepherding this review along—hopefully we can get this wrapped up "soon".

Like JOSS, JOSE reviews are not meant to be monolithic, requiring a response to both reviewers after the reviews are complete in the traditional manner.

Instead, recognizing the software-adjacent nature of the work, as reviewers provide comments and raise issues one-by-one, they can be addressed along the way, in a more continuous and conversational way in here.

(It's also easier for a second reviewer to just give a thumbs up if you have already addressed things raised here.)

So, please do work on addressing those issues if you can.
