### 2023-01-06 #143 @ThomasA
> > * **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support
>
> [ashleefv/ApplNumComp#4](https://github.com/ashleefv/ApplNumComp/issues/4)

I second this issue
