### 2023-01-06 #143 @ThomasA
Regarding the statement of need, I find it well described in the paper while it is not that clear in the repository documentation as such (looking at the README.md).
