### 2023-01-06 #143 @ThomasA
Regarding "telling the story", I think the paper does this somewhat, but the author's additional paper https://doi.org/10.18260/1-2--32072 referenced therein seems to do this even more so. I think it is fine that interested readers can go there for more information.
