### 2023-01-06 #143 @ThomasA
Aside from the few remarks above, I think this looks like an excellent and coherent collection of introductory educational material for the topics the module covers.
I apologise for the insanely long time I took to get around to completing this review.
