+++ L13 Parameter Estimation in Python.md
@@ -201,7 +201,9 @@ plt.show()
 * Solution [.py file](/CHEclassFa20/In%20Class%20Problem%20Solutions/Python/ODEParamEstimExample2.py)
 
 ## **Extended Example**
-* [Parameter estimation of ODE models for chemical kinetics](/CHEclassFa20/In%20Class%20Problem%20Activities/Python/ParameterEstimationOfODEModels.ipynb)
+* Parameter estimation of ODE models for chemical kinetics
+  - [Problem statement](https://github.com/ashleefv/ChESS2022/blob/master/5%20Interactive%20Coding%20Templates/J5_ParEstKinetics.ipynb)
+  - [Solution](https://github.com/ashleefv/ChESS2022/blob/master/5%20Interactive%20Coding%20Templates/J5_ParEstKinetics_solution.ipynb)
 
 ## **Reference for Further Exploration**
 * [Python curve fitting tools](https://docs.scipy.org/doc/scipy/reference/generated/scipy.optimize.curve_fit.html)
