### 2023-02-05 83d40db
Update L08 Python Functions.md
Updated from ChESS2022 extended example
