+++ L08 Python Functions.md
@@ -96,11 +96,15 @@ plt.show()
 ```
 ![Expected Graph](/Lesson_images/figure_L8.png)
 
+## **Extended Example**
+* Solving a system of ODEs
+  - [Problem statement](https://github.com/ashleefv/ChESS2022/blob/master/5%20Interactive%20Coding%20Templates/J4_NonisothermalPFR.ipynb)
+  - [Solution](https://github.com/ashleefv/ChESS2022/blob/master/5%20Interactive%20Coding%20Templates/M4_NonisothermalPFR_solution.mlx)
+
 ## **References for Further Exploration**
 * [Default values in Python](https://docs.python.org/3.7/tutorial/controlflow.html#more-on-defining-functions)
 * [Loops in Python](https://www.codementor.io/@sheena/python-generators-and-iterators-du1082iua). Python has more capabilities than MATLAB for loops using iterators and iterables.
 * [Python modules](https://docs.python.org/3/tutorial/modules.html)
-
 
 ## **Previous Lesson**
  * [L07 MATLAB Functions](/L07%20MATLAB%20Functions.md)
