### 2025-04-23 #143 @kyleniemeyer
@ashleefv this submission to JOSE is *very* stale (over two years!), but it looks like the only thing preventing publication was the missing community guidelines, which shouldn't take more than a few minutes to address.

If you still would like to publish this, please let us know by April 30, otherwise we'll close this out.
