### 2025-04-30 897633a
Update README.md
added statement of need adapted from paper.md
