+++ README.md
@@ -7,6 +7,9 @@ Note: this site is based on the Fall 2020 course offering: CHE 4753/5753 Applied
 Course repository:
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4711792.svg)](https://doi.org/10.5281/zenodo.4711792)
 
+## Statement of Need
+An introductory course on Applied Numerical Computing was developed for upper division undergraduate and graduate students interested in research and industrial projects involving mathematical modeling using computational tools. The class was designed to fill a gap distinct from introductory programming and numerical methods courses and was intended to be very practical. Graduate students new to computational research were the primary student group targeted for the course topics. We have previously detailed the course [FordVersypt 2019](https://doi.org/10.18260/1-2--32072) and the final course project [Ruggiero 2018]( https://doi.org/10.22369/issn.2153-4136/9/1/3). No single textbook covers all of the topics in the Applied Numerical Computing course. As the course emphasized best practices for reproducible research computing, we sought to apply the concept to the course materials for a reproducible computing educational experience. Here, we have curated a set of open educational resources for others to engage with the course lessons outside of the institution's learning management system and outside of the formal course. The main repository is organized into a set of lessons as Markdown-formatted text-based open-source webpages that include code snippets, explanatory text, reflection questions, sample solution software files, links to supplemental videos, related references, and other instructional materials. Each lesson can be considered as a stand alone module on a particular computational topic; however, there are conceptual links between some of the lessons. Many of the lessons include detailed MATLAB or Python coding examples, or both as in the case for Lesson 10: Python and MATLAB Plotting. For instance, Lesson 13: Parameter Estimation in Python provides a worked example for fitting parameters to one differential equation and the solution .py file followed by a second example for fitting parameters to multiple differential equations. Related Lessons 11 and 12 provide supportive general information and complimentary content with examples in MATLAB. Documentation associated with MATLAB or Python often only shows much simpler examples. Open-source worked examples are limited for even just solving multiple differential equations simultaneously independently from the parameter estimation. The full list of lessons is available [here](https://github.com/ashleefv/ApplNumComp#lessons). Other supporting materials include a list of [suggested readings](https://github.com/ashleefv/ApplNumComp/blob/master/RecommendedReading.md) with links to the sources, the full set of [computational assignments](https://github.com/ashleefv/ApplNumComp#computational-assignments), and the course [learning objectives](https://github.com/ashleefv/ApplNumComp#course-learning-objectives).
+
 ## Course Description
 Practical software tools for computational problem solving in science and engineering: version control (e.g., Git), mathematical typesetting (e.g., LaTeX), graphical user interfaces, and high level programming languages with libraries of solvers and visualization tools (e.g., Python and MATLAB). Application of numerical computing methods to solve systems of differential and algebraic equations and to estimate model parameters using optimization.
 
