### 2025-04-30 b9163f4
Update paper.md
trying to articulate that "Ford Versypt" is my surname
