### 2025-04-30 c803204
Create CONTRIBUTING.md
