### 2025-04-30 f5c782e
Update README.md
updated zenodo doi link for version 2.0
