### 2025-04-30 #143 @ashleefv
@kyleniemeyer @sconde @ThomasA > Sorry for losing track of this over time. I have added a community guidelines section to the README.md (and reiterated in a contributing.md file) and copied the statement of need from the paper.md into the README.md. After these updates, I've updated the code release reflecting commits since the original paper submission and updated the corresponding zenodo DOI.

I've also updated affiliation information.
