### 2025-05-01 #143 @kyleniemeyer
Thanks @ashleefv, can you just change the name of the Zenodo archive to match the paper? Should be able to update the metadata without a new release.
