### 2025-05-01 #143 @kyleniemeyer
@ashleefv also, just noticed that there are a few things near the end of the paper that might need to be updated, if you'd like ("next offering scheduled in 2022", etc.)
