+++ README.md
@@ -1,4 +1,4 @@
-# ApplNumComp: Applied Numerical Computing Course
+# ApplNumComp: An Open Access Introductory Course for Applied Numerical Computing
 
 This repository contains a set of lessons on Applied Numerical Computing covering Git for version control, LaTeX for typesetting, and MATLAB and Python for high-level programming and scientific computing. 
 
