+++ paper.bib
@@ -19,3 +19,13 @@
   title = {Building a MATLAB graphical user interface to solve ordinary differential equations as a final project for an interdisciplinary elective course on numerical computing},
   journal = {Journal of Computational Science Education}
 }
+
+@Article{Johns2023,
+  AUTHOR = {Johns, A. N. and Hesketh, R. P. and Stuber, M. D. and Ford Versypt, A. N.},
+  TITLE = {Numerical problem solving across the curriculum with Python and MATLAB using interactive coding templates: a workshop for chemical engineering faculty},
+  YEAR = {2023},
+  JOURNAL = {Proceedings of the ASEE Annual Conference},
+  doi = {10.18260/1-2-{}-43749},
+  url = {https://doi.org/10.18260/1-2-{}-43749},
+  City= {Baltimore, MD}
+}
