### 2025-05-01 a271a86
Update paper.md
updates to progress on the paper and the impacts of the repository since submission of the manuscript
