### 2025-05-01 #143 @kyleniemeyer
@ashleefv sorry, one final thing! Could you correct your coauthor's name on the Zenodo archive (write it out, I mean—it looks like a GitHub handle right now)? Otherwise, we are good to go.
