### 2025-05-01 #143 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
