+++ paper.bib
@@ -4,7 +4,7 @@
   YEAR = {2019},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
   doi = {10.18260/1-2-{}-32072},
-  url = {https://doi.org/10.18260/1-2-{}-32072},
+  url = {https://doi.org/10.18260/1-2--32072},
   City= {Tampa, FL}
 }
 
@@ -26,6 +26,6 @@
   YEAR = {2023},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
   doi = {10.18260/1-2-{}-43749},
-  url = {https://doi.org/10.18260/1-2-{}-43749},
+  url = {https://doi.org/10.18260/1-2--43749},
   City= {Baltimore, MD}
 }
