### 2025-05-01 6a23e84
Update paper.bib
fixing a special character in the urls for the bib
