### 2025-05-01 a7c9c56
Update paper.bib
still dealing with special character -- in the bib
