+++ paper.bib
@@ -3,8 +3,8 @@
   TITLE = {An interdisciplinary elective course to build computational skills for mathematical modeling in science and engineering},
   YEAR = {2019},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
-  doi = {10.18260/1-2-{}-32072},
-  url = {{https://doi.org/10.18260/1-2--32072}},
+  doi = {10.18260/1-2-\char`-32072},
+  url = {{https://doi.org/10.18260/1-2-\char`-32072}},
   City= {Tampa, FL}
 }
 
@@ -25,7 +25,7 @@
   TITLE = {Numerical problem solving across the curriculum with Python and MATLAB using interactive coding templates: a workshop for chemical engineering faculty},
   YEAR = {2023},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
-  doi = {10.18260/1-2-{}-43749},
-  url = {{https://doi.org/10.18260/1-2--43749}},
+  doi = {10.18260/1-2-\char`-43749},
+  url = {https://doi.org/10.18260/1-2-\char`-43749},
   City= {Baltimore, MD}
 }
