### 2025-05-01 #143 @ashleefv
I'm trying to deal with the special character -- in the doi for two refs in the bib. Hence, the many repeats of this as I debug:
