+++ paper.bib
@@ -3,14 +3,12 @@
   TITLE = {An interdisciplinary elective course to build computational skills for mathematical modeling in science and engineering},
   YEAR = {2019},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
-  doi = {10.18260/1-2-\char`-32072},
-  url = {{https://doi.org/10.18260/1-2-\char`-32072}},
+  doi = {{10.18260/1-2--32072}},
   City= {Tampa, FL}
 }
 
 @article{Ruggiero2018,
   doi = {10.22369/issn.2153-4136/9/1/3},
-  url = {https://doi.org/10.22369/issn.2153-4136/9/1/3},
   year = {2018},
   volume = {9},
   number = {1},
@@ -25,7 +23,6 @@
   TITLE = {Numerical problem solving across the curriculum with Python and MATLAB using interactive coding templates: a workshop for chemical engineering faculty},
   YEAR = {2023},
   JOURNAL = {Proceedings of the ASEE Annual Conference},
-  doi = {10.18260/1-2-\char`-43749},
-  url = {https://doi.org/10.18260/1-2-\char`-43749},
+  doi = {{10.18260/1-2--43749}},
   City= {Baltimore, MD}
 }
