### 2025-05-01 #143 @kyleniemeyer
I'm good with the contents of the paper, no edits needed from my perspective.
