### 2025-05-01 #143 @labarba
@openjournals/dev 👋  – it looks like this paper acceptance did not generate the final PDF... can you have a look?
