### 2025-05-01 269f25c
Create CITATION.cff
